+++ after/ftplugin/markdown.lua
@@ -1,4 +1,8 @@
 vim.opt_local.conceallevel = 2
-require('render-markdown').enable()
+vim.opt_local.tabstop = 2
+vim.opt_local.shiftwidth = 2
 
 vim.g.copilot_enabled = 0
+
+require('render-markdown').enable()
+
