+++ lua/plugins/treesitter/treesitter.lua
@@ -7,7 +7,9 @@ return {
             ensure_installed = { 'c', 'html', 'lua', 'markdown', 'vim', 'vimdoc' },
             -- Autoinstall languages that are not installed
             auto_install = true,
-            highlight = { enable = true },
+            highlight = { enable = true,
+                disable = { "diff" },
+            },
             indent = { enable = true },
         }
